+++ shaders/lib/Utility/TextureSampling.glsl
@@ -18,7 +18,35 @@ vec4 Cubic(float x)
     return w / 6.f;
 }
 
-vec4 BicubicTexture(in sampler2D tex, in vec2 coord, int resolution = 64)
+vec4 BicubicTexture(in sampler2D tex, in vec2 coord, int resolution = noiseTextureResolution)
+{
+
+	coord *= resolution;
+
+	float fx = fract(coord.x);
+    float fy = fract(coord.y);
+    coord.x -= fx;
+    coord.y -= fy;
+
+    vec4 xcubic = Cubic(fx);
+    vec4 ycubic = Cubic(fy);
+
+    vec4 c = vec4(coord.x - 0.5, coord.x + 1.5, coord.y - 0.5, coord.y + 1.5);
+    vec4 s = vec4(xcubic.x + xcubic.y, xcubic.z + xcubic.w, ycubic.x + ycubic.y, ycubic.z + ycubic.w);
+    vec4 offset = c + vec4(xcubic.y, xcubic.w, ycubic.y, ycubic.w) / s;
+
+    vec4 sample0 = texture2D(tex, vec2(offset.x, offset.z) / resolution);
+    vec4 sample1 = texture2D(tex, vec2(offset.y, offset.z) / resolution);
+    vec4 sample2 = texture2D(tex, vec2(offset.x, offset.w) / resolution);
+    vec4 sample3 = texture2D(tex, vec2(offset.y, offset.w) / resolution);
+
+    float sx = s.x / (s.x + s.y);
+    float sy = s.z / (s.z + s.w);
+
+    return mix( mix(sample3, sample2, sx), mix(sample1, sample0, sx), sy);
+}
+
+vec4 BicubicTexture(in sampler2D tex, in vec2 coord, ivec2 resolution)
 {
 
 	coord *= resolution;