+++ shaders/lib/VolumeRendering/Phase.glsl
@@ -14,7 +14,7 @@ float PhaseHenyeyGreenstein(in float cosTheta, in float g){
 }
 
 float PhaseMie(in float cosTheta) {
-    return PhaseHenyeyGreenstein(cosTheta, -0.974f);
+    return PhaseHenyeyGreenstein(cosTheta, -0.76f); // or 0.844 or 0.974
 }
 
 float PhaseRayleigh(in vec3 v, in vec3 l){